// src/services/api/@types/transaction.ts
export type ItemType = 'package' | 'product' | 'freeze'
export type DurationType = 'day' | 'month' | 'year'
export type DiscountType = 'percent' | 'nominal'
export type PaymentStatus = 0 | 1 | 2 | 3 // 0 = belum bayar, 1 = bayar semua, 2 = bayar sebagian. 3 = bayar sebagian dan aktifkan member package jika terdapat package
export type PackageType = 'membership' | 'pt_program' | 'class'

export interface TransactionItem {
  item_type: ItemType
  package_id?: number
  product_id?: number
  name: string
  quantity: number
  price: number
  sell_price: number
  discount_type?: DiscountType | null
  discount?: number | null
  duration?: number | null
  duration_type?: DurationType | null
  session_duration?: number | null
  extra_session?: number | null
  extra_day?: number | null
  start_date?: Date | null
  notes?: string | null

  // custom fields
  is_promo?: number
  loyalty_point?: number
  package_type?: PackageType | null
}

interface Payment {
  amount: number
  id: number
  //   reference_no: string
  //   notes?: string
  //   payment_date: string
}

interface RefundItem {
  id: number
  amount: number
  //   notes?: string
}

export interface CheckoutRequest {
  club_id: number
  member_id?: number
  discount_type: DiscountType
  discount?: number
  tax_rate?: number
  is_paid: PaymentStatus
  due_date?: string
  notes?: string
  items: TransactionItem[]
  payments: Payment[]
  refund_from: RefundItem[]
}